requirements.txt
Flask==2.3.2
transformers==4.28.1
torch==2.0.1
PyPDF2==3.0.1
python-docx==0.8.11